requirements.txt
# --allow-all-external

Django==3.2.13
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-rest-knox==4.2.0
